migrate append_job_logs_to_replica_log to serve_utils
