+++ sky/serve/replica_managers.py
@@ -168,24 +168,6 @@ def terminate_cluster(cluster_name: str,
             time.sleep(gap_seconds)
 
 
-def append_job_logs_to_replica_log(new_replica_log_file: str,
-                                   job_log_file_name: Optional[str],
-                                   replica_id: int) -> None:
-
-    if job_log_file_name is not None:
-        logger.info(f'\n== End of logs (Replica: {replica_id}) ==')
-        with open(new_replica_log_file, 'a',
-                  encoding='utf-8') as replica_log_file, open(
-                      job_log_file_name, 'r', encoding='utf-8') as job_file:
-            replica_log_file.write(job_file.read())
-        os.remove(job_log_file_name)
-    else:
-        with open(new_replica_log_file, 'a',
-                  encoding='utf-8') as replica_log_file:
-            replica_log_file.write('Failed to sync down job logs from replica '
-                                   f'{replica_id}.\n')
-
-
 def _get_resources_ports(task_yaml: str) -> str:
     """Get the resources ports used by the task."""
     task = sky.Task.from_yaml(task_yaml)
@@ -756,8 +738,9 @@ class SkyPilotReplicaManager(ReplicaManager):
                 controller_utils.download_and_stream_latest_job_log(
                     backend, handle, replica_job_logs_dir))
 
-            append_job_logs_to_replica_log(new_replica_log_file,
-                                           job_log_file_name, replica_id)
+            serve_utils.append_job_logs_to_replica_log(new_replica_log_file,
+                                                       job_log_file_name,
+                                                       replica_id)
 
         logger.info(f'Terminating replica {replica_id}...')
         info = serve_state.get_replica_info_from_id(self._service_name,
